hmf: hmf-pipeline: Propagate grid-engine-core for a regular 'qsub' command.
The regexp in jobNative() expects "qsub" and not an absolute path to the
qsub command.  We can either change the regexp, or propagate grid-engine-core.
In this case, we want to stay as close as possible to the upstream code, so
instead of patching the jobNative() function, we will make sure the qsub
command is available by propagating grid-engine-core.

* umcu/packages/hmf.scm (hmf-pipeline): Remove patching of the qsub command.