guix: guixr: Fix race condition in symbolic link creation.
The “guixr” command automatically creates a symbolic link to point to the
latest version of Guix.  If multiple invocations of “guixr” attempt to
create this link, one succeeds and the rest receives an error (File exists).

In this patch we ignore the error from “ln”.  Fun fact: The lines involved are
automatically marked with a ||: emoticon.  It's as if the creators of the null
utility were aware of its expressiveness.

* umcu/packages/guix.scm (guixr): Ignore “ln” error output; Bump version number.